+++ app/src/main/java/com/personal/happygallery/app/booking/BookingCancelService.java
@@ -17,6 +17,7 @@ import com.personal.happygallery.infra.booking.SlotRepository;
 import com.personal.happygallery.infra.pass.PassLedgerRepository;
 import com.personal.happygallery.infra.pass.PassPurchaseRepository;
 import java.time.Clock;
+import java.util.function.BiConsumer;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -53,90 +54,41 @@ public class BookingCancelService {
      *
      * <ol>
      *   <li>access_token으로 예약 조회 및 검증</li>
-     *   <li>BOOKED 상태 확인</li>
-     *   <li>슬롯 반납: 비관적 락 + booked_count--</li>
-     *   <li>CANCELED 이력 저장</li>
-     *   <li>8회권: D-1 이전이면 REFUND ledger(+1) + refundCredit(). 이후면 크레딧 소멸.</li>
-     *   <li>예약금: D-1 이전이면 PG 환불 요청(Refund REQUESTED).</li>
-     *   <li>booking.cancel() + save</li>
+     *   <li>상태 확인 → 슬롯 반납 → 이력 → 환불 → 취소 → 알림</li>
      * </ol>
      *
      * @return (취소된 booking, 환불 가능 여부)
      */
     public CancelResult cancelBooking(Long bookingId, String accessToken) {
-
-        // 1. 예약 로드 (accessToken 검증)
         Booking booking = bookingSupport.findByToken(bookingId, accessToken);
-
-        // 2. 상태 체크 — BOOKED 상태만 취소 가능
-        if (booking.getStatus() != BookingStatus.BOOKED) {
-            throw new HappyGalleryException(ErrorCode.INVALID_INPUT, "취소할 수 없는 예약 상태입니다.");
-        }
-
-        // 3. 슬롯 반납 — 비관적 락 + booked_count--
-        Slot slot = slotRepository.findByIdWithLock(booking.getSlot().getId())
-                .orElseThrow(() -> new NotFoundException("슬롯"));
-        slot.decrementBookedCount();
-        slotRepository.save(slot);
-
-        // 4. CANCELED 이력 저장 (append-only)
-        bookingSupport.recordHistory(booking, BookingHistoryAction.CANCELED, slot, null, "CUSTOMER", null);
-
-        // 5. 환불 가능 여부 판단 (D-1 00:00 Asia/Seoul 기준)
-        boolean refundable = TimeBoundary.isRefundable(slot.getStartAt(), clock);
-
-        if (booking.isPassBooking()) {
-            // 5a. 8회권 결제 취소 — D-1 이전이면 크레딧 복구, 이후면 소멸 유지
-            if (refundable) {
-                PassPurchase pass = booking.getPassPurchase();
-                passLedgerRepository.save(
-                        new PassLedger(pass, PassLedgerType.REFUND, 1, booking.getId()));
-                pass.refundCredit();
-                passPurchaseRepository.save(pass);
-            }
-        } else {
-            // 5b. 예약금 결제 취소 — D-1 이전이면 PG 환불 요청
-            if (refundable) {
-                refundExecutionService.processBookingRefund(booking.getId(), booking.getDepositAmount());
-            }
-        }
-
-        // 6. 예약 취소 처리
-        booking.cancel();
-        bookingRepository.save(booking);
-
-        // 7. 취소 알림 (+ 예약금 환불 시 환불 알림)
-        bookingSupport.notifyBookingGuest(booking, NotificationEventType.BOOKING_CANCELED);
-        if (refundable && !booking.isPassBooking()) {
-            bookingSupport.notifyBookingGuest(booking, NotificationEventType.DEPOSIT_REFUNDED);
-        }
-
-        return new CancelResult(booking, refundable);
+        return cancelInternal(booking, bookingSupport::notifyBookingGuest);
     }
 
     /**
      * 회원 예약을 취소한다. accessToken 대신 userId 소유권으로 검증한다.
      */
     public CancelResult cancelMemberBooking(Long bookingId, Long userId) {
-
-        // 1. 예약 로드 (userId 소유권 검증)
         Booking booking = bookingSupport.findByIdAndUserId(bookingId, userId);
+        return cancelInternal(booking, bookingSupport::notifyBookingUser);
+    }
 
-        // 2. 상태 체크 — BOOKED 상태만 취소 가능
+    private CancelResult cancelInternal(Booking booking,
+                                         BiConsumer<Booking, NotificationEventType> notify) {
+        // 1. 상태 체크 — BOOKED 상태만 취소 가능
         if (booking.getStatus() != BookingStatus.BOOKED) {
             throw new HappyGalleryException(ErrorCode.INVALID_INPUT, "취소할 수 없는 예약 상태입니다.");
         }
 
-        // 3. 슬롯 반납
+        // 2. 슬롯 반납 — 비관적 락 + booked_count--
         Slot slot = slotRepository.findByIdWithLock(booking.getSlot().getId())
                 .orElseThrow(() -> new NotFoundException("슬롯"));
         slot.decrementBookedCount();
         slotRepository.save(slot);
 
-        // 4. CANCELED 이력 저장
+        // 3. CANCELED 이력 저장 (append-only)
         bookingSupport.recordHistory(booking, BookingHistoryAction.CANCELED, slot, null, "CUSTOMER", null);
 
-        // 5. 환불 가능 여부 판단
+        // 4. 환불 가능 여부 판단 (D-1 00:00 Asia/Seoul 기준)
         boolean refundable = TimeBoundary.isRefundable(slot.getStartAt(), clock);
 
         if (booking.isPassBooking()) {
@@ -153,14 +105,14 @@ public class BookingCancelService {
             }
         }
 
-        // 6. 예약 취소 처리
+        // 5. 예약 취소 처리
         booking.cancel();
         bookingRepository.save(booking);
 
-        // 7. 취소 알림
-        bookingSupport.notifyBookingUser(booking, NotificationEventType.BOOKING_CANCELED);
+        // 6. 취소 알림 (+ 예약금 환불 시 환불 알림)
+        notify.accept(booking, NotificationEventType.BOOKING_CANCELED);
         if (refundable && !booking.isPassBooking()) {
-            bookingSupport.notifyBookingUser(booking, NotificationEventType.DEPOSIT_REFUNDED);
+            notify.accept(booking, NotificationEventType.DEPOSIT_REFUNDED);
         }
 
         return new CancelResult(booking, refundable);
